Add basic NPC movement to Scene1
