--- My project/Assets/CircleMovement.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 5f84226c5726d1a43befa5f1588e6ccf
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
