+++ My project/Assets/Scripts/CircleMovement.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 5cd1e9811ef82164198873144c6da4e5
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
