--- My project/Assets/Scripts/HorizontalMovement.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: cc2367edfc93a16499d99b18f1033454
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
