+++ My project/Assets/Scripts/Trigger.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: ff10b1579498192499ecbdb50ad2fba4
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
