+++ My project/Assets/TriggerPoint.cs
@@ -1,0 +1,22 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class TriggerPoint : MonoBehaviour
+{
+    public Transform t;
+    public bool isJump;
+    public int waitTime;
+
+    // Start is called before the first frame update
+    void Start()
+    {
+        
+    }
+
+    // Update is called once per frame
+    void Update()
+    {
+        
+    }
+}
