+++ My project/Assets/TriggerPoint.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 13f1d7eeace3b0844937b4ef0a72efba
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
