+++ My project/Assets/Unity.VisualScripting.Generated.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 69567fecdf9d3d0449250e8321662a5b
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
