+++ My project/Assets/Unity.VisualScripting.Generated/VisualScripting.Flow.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 6a7b128aab7b8d14eb7cc96f4d83edda
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
