+++ My project/Assets/Unity.VisualScripting.Generated/VisualScripting.Flow/UnitOptions.db.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: b0f17f22c7923b847a95274f15300846
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
