--- My project/Assets/triggerPoop.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: bfd7a68239cb5eb4387358f16412a6a1
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
